fix(testsuite): change locations of resource_test_suite
