+++ internal/resource/test_helpers.go
@@ -10,7 +10,7 @@ MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
 See the Mulan PSL v2 for more details.
 */
 
-package controller
+package resource
 
 import (
 	"fmt"
