+++ make/development.mk
@@ -21,14 +21,31 @@ test: manifests generate fmt vet envtest ## Run tests.
 	TEST_USE_EXISTING_CLUSTER=true TELEMETRY_REPORT_HOST=http://openwebapi.test.alipay.net \
 	DISABLE_TELEMETRY=true LOG_VERBOSITY=0 \
 	KUBEBUILDER_ASSETS="$(shell $(ENVTEST) use $(ENVTEST_K8S_VERSION) --bin-dir $(LOCALBIN) -p path)" \
-	ginkgo run --coverprofile=cover.profile --cpuprofile=cpu.profile --memprofile=mem.profile --cover \
-	--output-dir=testreports --keep-going --json-report=report.json --label-filter='$(CASE_LABEL_FILTERS)' ./...
+	go run github.com/onsi/ginkgo/v2/ginkgo -r --coverprofile=cover.profile --cpuprofile=cpu.profile --memprofile=mem.profile --cover \
+	--output-dir=testreports --keep-going --json-report=report.json --label-filter='!long-run'
+	
+.PHONY: test-all
+test-all: manifests generate fmt vet envtest ## Run tests to get cpu and mem profile
+	TEST_USE_EXISTING_CLUSTER=true TELEMETRY_REPORT_HOST=http://openwebapi.test.alipay.net \
+	DISABLE_TELEMETRY=true LOG_VERBOSITY=0 \
+	KUBEBUILDER_ASSETS="$(shell $(ENVTEST) use $(ENVTEST_K8S_VERSION) --bin-dir $(LOCALBIN) -p path)" \
+	go run github.com/onsi/ginkgo/v2/ginkgo -r --coverprofile=cover.profile --cpuprofile=cpu.profile --memprofile=mem.profile --cover \
+	--output-dir=testreports --keep-going --json-report=report.json --label-filter='$(CASE_LABEL_FILTERS)'
 
-.PHONY: testreport
-testreport: test ## Generate test reports
+REPORT_PORT ?= 8480
+
+.PHONY: coverage
+coverage: test ## Generate test reports
 	@echo "generating test reports..."
-	@go tool cover -html=testreports/cover.profile -o testreports/cover.html
-	@cd testreports && python3 -m http.server --bind 0.0.0.0 8480
+	@go tool cover -html=testreports/cover.out -o testreports/cover.html
+	@cd testreports && python3 -m http.server --bind 0.0.0.0 $(REPORT_PORT)
+
+.PHONY: run-coverage
+run-coverage:  ## Generate integration test coverage report.
+	@go tool covdata textfmt -i=testreports/covdata -o=testreports/covdata.txt
+	@go tool cover -html=testreports/covdata.txt -o testreports/integration.html
+	@go tool cover -html=testreports/cover.out -o testreports/unit.html
+	@cd testreports && python3 -m http.server --bind 0.0.0.0 $(REPORT_PORT)
 
 .PHONY: GOLANGCI_LINT
 GOLANGCI_LINT ?= $(LOCALBIN)/golangci-lint
